third_party/WebKit/LayoutTests/platform/mac/fast/spatial-navigation/snav-tiny-table-traversal-expected.txt

a
a	a	a
a	a	a
a	a	a
a
a
PASS successfullyParsed is true

TEST COMPLETE
PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "4"
PASS gFocusedDocument.activeElement.getAttribute("id") is "7"
PASS gFocusedDocument.activeElement.getAttribute("id") is "end"
PASS gFocusedDocument.activeElement.getAttribute("id") is "7"
PASS gFocusedDocument.activeElement.getAttribute("id") is "8"
PASS gFocusedDocument.activeElement.getAttribute("id") is "9"
PASS gFocusedDocument.activeElement.getAttribute("id") is "6"
PASS gFocusedDocument.activeElement.getAttribute("id") is "right"
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 6. Was 3.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 3. Was start.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 2. Was start.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 1. Was start.
